# backend/api/openapi.yaml
openapi: "3.1.0"
info: {title: Oh-My-AIHub API, version: "0.1.0", description: "Oh-My-AIHub 前后端契约的唯一来源（见 ADR）。覆盖全部 /api/** JSON 接口；/v1/**、/v1beta/** 外部模型协议入口只登记路径与认证。\n\n约定：金额一律为十进制字符串；时间为 RFC 3339；错误统一为 {\"error\":{\"code\",\"message\"}}；除 GET/HEAD/OPTIONS 外的 /api 请求须携带与 Host 同源的 Origin。\n\n扩展字段 x-access 声明路由门禁，由契约测试与路由表逐项对照：public 无需会话；session 需会话（允许首次改密前访问）；ready 需会话且已完成首次改密；admin 在 ready 基础上需管理员；gateway_key 凭网关 API Key。"}
servers:
  - url: /
    description: "同源部署：开发环境由 Vite 代理，Compose 环境由 Nginx 代理。"
...
paths:
...
  /api/c2c/market:
    get:
      operationId: getC2CMarket
      tags: [c2c]
      summary: "C2C 市场"
      x-access: ready
      security:
        - sessionCookie: []
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2CMarket"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
  /api/c2c/orders:
    post:
      operationId: createC2COrder
      tags: [c2c]
      summary: "创建 C2C 订单"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Idempotency-Key"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/C2CCreateOrderRequest"}
          multipart/form-data:
            schema:
              type: object
              description: "payload 为 C2CCreateOrderRequest 的 JSON 字符串；其余部分为各收款方式 qr_field 指向的图片文件。"
              properties:
                payload: {type: string}
              required: [payload]
              additionalProperties: {type: string, format: binary}
      responses:
        '201':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2COrderEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/c2c/orders/{orderID}:
    get:
      operationId: getC2COrder
      tags: [c2c]
      summary: "订单详情"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/orderID"
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2COrderEnvelope"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
  /api/c2c/orders/{orderID}/payment-methods/{methodID}/qr:
    get:
      operationId: getC2CPaymentQR
      tags: [c2c]
      summary: "收款码图片"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/orderID"
        - $ref: "#/components/parameters/methodID"
      responses:
        '200':
          description: "已清洗的收款码图片，Content-Type 随图片格式。"
          content:
            image/png:
              schema: {type: string, format: binary}
            image/jpeg:
              schema: {type: string, format: binary}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
  /api/c2c/orders/{orderID}/take:
    post:
      operationId: takeC2COrder
      tags: [c2c]
      summary: "吃单"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/orderID"
        - $ref: "#/components/parameters/Idempotency-Key"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/C2CTakeOrderRequest"}
      responses:
        '201':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2CTradeEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/c2c/orders/{orderID}/cancel:
    post:
      operationId: cancelC2COrder
      tags: [c2c]
      summary: "撤销订单"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/orderID"
        - $ref: "#/components/parameters/Idempotency-Key"
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2COrderEnvelope"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
  /api/c2c/me:
    get:
      operationId: getC2CMyActivity
      tags: [c2c]
      summary: "我的订单与交易"
      x-access: ready
      security:
        - sessionCookie: []
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2CMyActivity"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
  /api/c2c/trades/{tradeID}:
    get:
      operationId: getC2CTrade
      tags: [c2c]
      summary: "交易详情"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/tradeID"
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2CTradeViewEnvelope"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
  /api/c2c/trades/{tradeID}/paid:
    post:
      operationId: markC2CTradePaid
      tags: [c2c]
      summary: "标记已付款"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/tradeID"
        - $ref: "#/components/parameters/Idempotency-Key"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/C2CMarkPaidRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2CTradeEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/c2c/trades/{tradeID}/cancel:
    post:
      operationId: cancelC2CTrade
      tags: [c2c]
      summary: "取消交易"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/tradeID"
        - $ref: "#/components/parameters/Idempotency-Key"
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2CTradeEnvelope"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
  /api/c2c/trades/{tradeID}/release:
    post:
      operationId: confirmC2CReceipt
      tags: [c2c]
      summary: "确认收款并放行"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/tradeID"
        - $ref: "#/components/parameters/Idempotency-Key"
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2CTradeEnvelope"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
  /api/c2c/trades/{tradeID}/dispute:
    post:
      operationId: openC2CDispute
      tags: [c2c]
      summary: "发起争议"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/tradeID"
        - $ref: "#/components/parameters/Idempotency-Key"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/C2CStatementRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2CTradeEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/c2c/trades/{tradeID}/statements:
    post:
      operationId: addC2CStatement
      tags: [c2c]
      summary: "补充争议陈述"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/tradeID"
        - $ref: "#/components/parameters/Idempotency-Key"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/C2CStatementRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2CTradeEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/admin/c2c/disputes:
    get:
      operationId: listC2CDisputes
      tags: [admin-c2c]
      summary: "管理员：争议列表"
      x-access: admin
      security:
        - sessionCookie: []
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2CAdminTradeList"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
  /api/admin/c2c/orders/{orderID}/cancel:
    post:
      operationId: adminCancelC2COrder
      tags: [admin-c2c]
      summary: "管理员：撤销订单"
      x-access: admin
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/orderID"
        - $ref: "#/components/parameters/Idempotency-Key"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/C2CAdminCancelOrderRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2COrderEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/admin/c2c/trades/{tradeID}/resolve:
    post:
      operationId: resolveC2CDispute
      tags: [admin-c2c]
      summary: "管理员：裁决争议"
      x-access: admin
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/tradeID"
        - $ref: "#/components/parameters/Idempotency-Key"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/C2CResolveRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2CAdminTradeEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: oma_session
      description: "登录后由服务端设置的 HttpOnly 会话 Cookie。"
    bearerKey: {type: http, scheme: bearer, description: "平台 API Key：Authorization: Bearer <key>（OpenAI 协议）。"}
    anthropicKey:
      type: apiKey
      in: header
      name: x-api-key
      description: "平台 API Key（Anthropic 协议）。"
    geminiKey:
      type: apiKey
      in: header
      name: x-goog-api-key
      description: "平台 API Key（Gemini 协议）。"
  parameters:
...
    orderID:
      name: orderID
      in: path
      required: true
      description: "C2C 订单 ID。"
      schema: {type: string}
    tradeID:
      name: tradeID
      in: path
      required: true
      description: "C2C 交易 ID。"
      schema: {type: string}
    methodID:
      name: methodID
      in: path
      required: true
      description: "收款方式 ID。"
      schema: {type: string}
...
    Idempotency-Key:
      name: Idempotency-Key
      in: header
      required: false
      description: "幂等键；相同键重复提交返回同一结果。"
      schema: {type: string}
...
  responses:
    BadRequest:
      description: "请求格式无效或查询参数非法（invalid_json、invalid_request、invalid_input）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Unauthorized:
      description: "未登录或会话失效（authentication_required）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Forbidden:
      description: "被拒绝：跨站请求（cross_origin_request）、首次登录须改密（password_change_required）、需要管理员（administrator_required）或无权操作（forbidden）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    NotFound:
      description: "资源不存在（not_found）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Conflict:
      description: "版本或状态冲突（conflict、ledger_conflict、c2c_conflict、fee_rate_conflict 等）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Unprocessable:
      description: "业务校验失败（invalid_input、insufficient_spendable_capacity、unsafe_upstream 等）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
...
  schemas:
    Amount: {type: string, pattern: "^-?[0-9]+(\\.[0-9]{1,9})?$", description: "积分金额的十进制字符串，最多 9 位小数，不使用浮点数。"}
...
    Error:
      type: object
      properties:
        code: {type: string}
        message: {type: string}
      required: [code, message]
      additionalProperties: false
    ErrorResponse:
      type: object
      description: "所有 /api 错误的统一信封；code 为稳定的机器可读标识，message 为中文提示。"
      properties:
        error: {$ref: "#/components/schemas/Error"}
      required: [error]
      additionalProperties: false
...
    C2COrderEnvelope:
      type: object
      properties:
        order: {$ref: "#/components/schemas/C2COrder"}
      required: [order]
      additionalProperties: false
...
    C2CTradeEnvelope:
      type: object
      properties:
        trade: {$ref: "#/components/schemas/C2CTrade"}
      required: [trade]
      additionalProperties: false
    C2CAdminTradeEnvelope:
      type: object
      properties:
        trade: {$ref: "#/components/schemas/C2CAdminTrade"}
      required: [trade]
      additionalProperties: false
    C2CTradeViewEnvelope:
      type: object
      description: "管理员得到 C2CAdminTrade，其他账户得到 C2CTrade。"
      properties:
        trade:
          oneOf:
            - $ref: "#/components/schemas/C2CTrade"
            - $ref: "#/components/schemas/C2CAdminTrade"
      required: [trade]
      additionalProperties: false
    C2CAdminTradeList:
      type: object
      properties:
        trades:
          type: array
          items: {$ref: "#/components/schemas/C2CAdminTrade"}
      required: [trades]
      additionalProperties: false
    C2CMarket:
      type: object
      properties:
        metrics:
          type: object
          properties:
            guidance_price_fen: {type: integer}
            latest_price_fen: {type: [integer, "null"]}
            best_ask_fen: {type: [integer, "null"]}
          required: [guidance_price_fen, latest_price_fen, best_ask_fen]
          additionalProperties: false
        sell_orders:
          type: array
          items: {$ref: "#/components/schemas/C2COrder"}
      required: [metrics, sell_orders]
      additionalProperties: false
    C2CMyActivity:
      type: object
      properties:
        orders:
          type: array
          items: {$ref: "#/components/schemas/C2COrder"}
        trades:
          type: array
          items: {$ref: "#/components/schemas/C2CTrade"}
      required: [orders, trades]
      additionalProperties: false
...
    C2CCreateOrderRequest:
      type: object
      properties:
        unit_price_fen: {type: integer}
        total: {$ref: "#/components/schemas/Amount"}
        minimum: {$ref: "#/components/schemas/Amount"}
        maximum: {$ref: "#/components/schemas/Amount"}
        payment_methods:
          type: array
          items: {$ref: "#/components/schemas/C2CPaymentMethodRequest"}
      required: [unit_price_fen, total, minimum, maximum, payment_methods]
      additionalProperties: false
    C2CTakeOrderRequest:
      type: object
      properties:
        quantity: {$ref: "#/components/schemas/Amount"}
        payment_method_id: {type: string}
      required: [quantity, payment_method_id]
      additionalProperties: false
    C2CMarkPaidRequest:
      type: object
      properties:
        payment_reference: {type: string}
      required: [payment_reference]
      additionalProperties: false
    C2CStatementRequest:
      type: object
      properties:
        statement: {type: string}
      required: [statement]
      additionalProperties: false
    C2CAdminCancelOrderRequest:
      type: object
      properties:
        reason: {type: string}
      required: [reason]
      additionalProperties: false
    C2CResolveRequest:
      type: object
      properties:
        action: {type: string, enum: [release_to_buyer, return_to_seller, extend_review, restrict_buyer, restrict_seller]}
        reason: {type: string}
      required: [action, reason]
      additionalProperties: false
...